Split plugin test classes and share params
This should increase test parallelization a bit more and clarify test grouping.
