+++ plugin/src/integrationTest/kotlin/com/vanniktech/maven/publish/SpecialCasePluginTest.kt
@@ -5,27 +5,9 @@ import com.google.testing.junit.testparameterinjector.junit5.TestParameterInject
 import com.vanniktech.maven.publish.ProjectResultSubject.Companion.assertThat
 import com.vanniktech.maven.publish.TestOptions.Signing.GPG_KEY
 import com.vanniktech.maven.publish.TestOptions.Signing.NO_SIGNING
-import java.nio.file.Path
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.io.TempDir
 
-class MavenPublishPluginSpecialCaseTest {
-  @TempDir
-  lateinit var testProjectDir: Path
-
-  @TestParameter(valuesProvider = TestOptionsConfigProvider::class)
-  lateinit var config: TestOptions.Config
-
-  @TestParameter(valuesProvider = GradleVersionProvider::class)
-  lateinit var gradleVersion: GradleVersion
-
-  private val testOptions
-    get() = TestOptions(config, NO_SIGNING, gradleVersion)
-
-  @BeforeEach
-  fun setup() {
-    gradleVersion.assumeSupportedJdkVersion()
-  }
+class SpecialCasePluginTest : BasePluginTest() {
+  override val testOptions get() = TestOptions(config, NO_SIGNING, gradleVersion)
 
   @TestParameterInjectorTest
   fun artifactIdThatContainsProjectNameProducesCorrectArtifactId(
@@ -198,7 +180,7 @@ class MavenPublishPluginSpecialCaseTest {
   @TestParameterInjectorTest
   fun withoutSigning() {
     val project = javaProjectSpec()
-    val result = project.run(fixtures, testProjectDir, testOptions.copy(signing = NO_SIGNING))
+    val result = project.run(fixtures, testProjectDir, testOptions)
 
     assertThat(result).outcome().succeeded()
     assertThat(result).artifact("jar").exists()
@@ -231,68 +213,5 @@ class MavenPublishPluginSpecialCaseTest {
     assertThat(result).sourcesJar().containsAllSourceFiles()
     assertThat(result).javadocJar().exists()
     assertThat(result).javadocJar().isSigned()
-  }
-
-  @TestParameterInjectorTest
-  fun dokka() {
-    val kotlinVersion = KotlinVersion.values().last()
-    val original = kotlinJvmProjectSpec(kotlinVersion)
-    val project = original.copy(
-      plugins = original.plugins + dokkaPlugin,
-      basePluginConfig = original.basePluginConfig.replace("JavadocJar.Empty()", "JavadocJar.Dokka(\"dokkaHtml\")"),
-    )
-    val result = project.run(fixtures, testProjectDir, testOptions)
-
-    assertThat(result).outcome().succeeded()
-    assertThat(result).artifact("jar").exists()
-    assertThat(result).pom().exists()
-    assertThat(result).pom().matchesExpectedPom(kotlinStdlibJdk(kotlinVersion))
-    assertThat(result).module().exists()
-    assertThat(result).sourcesJar().exists()
-    assertThat(result).sourcesJar().containsAllSourceFiles()
-    assertThat(result).javadocJar().exists()
-    assertThat(result).javadocJar().containsFiles(ignoreAdditionalFiles = true, "index.html")
-  }
-
-  @TestParameterInjectorTest
-  fun dokka2() {
-    val kotlinVersion = KotlinVersion.values().last()
-    val original = kotlinJvmProjectSpec(kotlinVersion)
-    val project = original.copy(
-      plugins = original.plugins + dokka2Plugin,
-      basePluginConfig = original.basePluginConfig.replace("JavadocJar.Empty()", "JavadocJar.Dokka(\"dokkaGeneratePublicationHtml\")"),
-    )
-    val result = project.run(fixtures, testProjectDir, testOptions)
-
-    assertThat(result).outcome().succeeded()
-    assertThat(result).artifact("jar").exists()
-    assertThat(result).pom().exists()
-    assertThat(result).pom().matchesExpectedPom(kotlinStdlibJdk(kotlinVersion))
-    assertThat(result).module().exists()
-    assertThat(result).sourcesJar().exists()
-    assertThat(result).sourcesJar().containsAllSourceFiles()
-    assertThat(result).javadocJar().exists()
-    assertThat(result).javadocJar().containsFiles(ignoreAdditionalFiles = true, "index.html")
-  }
-
-  @TestParameterInjectorTest
-  fun dokka2Javadoc() {
-    val kotlinVersion = KotlinVersion.values().last()
-    val original = kotlinJvmProjectSpec(kotlinVersion)
-    val project = original.copy(
-      plugins = original.plugins + dokka2JavadocPlugin,
-      basePluginConfig = original.basePluginConfig.replace("JavadocJar.Empty()", "JavadocJar.Dokka(\"dokkaGeneratePublicationJavadoc\")"),
-    )
-    val result = project.run(fixtures, testProjectDir, testOptions)
-
-    assertThat(result).outcome().succeeded()
-    assertThat(result).artifact("jar").exists()
-    assertThat(result).pom().exists()
-    assertThat(result).pom().matchesExpectedPom(kotlinStdlibJdk(kotlinVersion))
-    assertThat(result).module().exists()
-    assertThat(result).sourcesJar().exists()
-    assertThat(result).sourcesJar().containsAllSourceFiles()
-    assertThat(result).javadocJar().exists()
-    assertThat(result).javadocJar().containsFiles(ignoreAdditionalFiles = true, "index.html")
   }
 }
